requirements.txt
streamlit
python-dotenv
pydantic
openai
requests
geopy